+++ src/main/java/gr/knowledge/internship/introduction/repository/BonusRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import gr.knowledge.internship.introduction.entity.Bonus;
 
-public interface BonusRepository extends JpaRepository<Bonus, Integer>{
+public interface BonusRepository extends JpaRepository<Bonus, Long>{
 
 }
